requirements.txt
Pillow==9.5.0
openai==0.27.0
boto3==1.17.4
